% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuddleMain.R
\name{OneHot2Label}
\alias{OneHot2Label}
\title{Obtaining Labels}
\usage{
OneHot2Label(OHE, Label)
}
\arguments{
\item{OHE}{an r-by-n one-hot encoding matrix.}

\item{Label}{an r-by-1 vector of values or levels which a label can take.}
}
\value{
An n-by-1 vector of labels.
}
\description{
Convert a one-hot encoding matrix to a vector of labels.
}
\seealso{
CheckNonNumeric(), GetPrecision(), FetchBuddle(), MakeConfusionMatrix(), Split2TrainTest(), TrainBuddle()
}
